backend/requirements.txt
aiohttp==4.0.0a1
aiohttp-cors==0.7.0
aiohttp-session==2.9.0
aiopg==1.0.0
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.9
jwt==0.6.1
multidict==4.7.5
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycparser==2.20
six==1.14.0
typing-extensions==3.7.4.2
yarl==1.4.2
